x86_64/x32: Fix various types issues
* __errno.c/errno.h - use correct pointer types.
* stat.h - Add 64-bit stat structure mapping 64-bit stat syscall.
* time.h - Fix timezone type to avoid redefinition.

+++ libc/bionic/__errno.c
@@ -28,7 +28,7 @@
 #include <errno.h>
 #include <bionic_tls.h>
 
-volatile int*  __errno( void )
+volatile intptr_t*  __errno( void )
 {
-  return  &((volatile int*)__get_tls())[TLS_SLOT_ERRNO];
+  return  &((volatile intptr_t *)__get_tls())[TLS_SLOT_ERRNO];
 }
